feat(btree): composite keys, so an index can hold a relation not just a map
The topology's structures are not key->value. Collection membership is
(collection, node) and adjacency is (node, edge); asking "every member of
this collection" or "every edge from this node" has to be a contiguous run
in one index, not a separate structure per relation. A u64 key cannot hold
a pair, so the tree took one and the topology would have needed its own.

Keys are u128 now. Two u64s pack into one, the high half sorts first, and a
prefix scan falls out of an ordinary range: range(compose(owner, 0),
compose(owner, u64::MAX)). `range` gained tests against a BTreeMap oracle
over eight awkward windows, and a composite test that drops one owner
entirely and checks its neighbours are undisturbed.

The cost is real and named: an entry is 24 bytes rather than 16, so a 4 KB
page holds 170 keys instead of 255. Measured on 50k random keys the tree
now uses 518 pages at 97 entries per page - the depth is unchanged at three
levels, so lookups still cost the same three page reads. The shallow-tree
test asserted a fanout bound that only held for the narrower key; it now
states where its number comes from.

+++ src/storage/pagedstore.rs
@@ -61,7 +61,7 @@ impl PagedStore {
     ///
     /// The old version's space is returned *after* the new one is written, so a
     /// failure part-way leaves the previous record intact rather than neither.
-    pub(crate) fn put(&mut self, key: u64, bytes: &[u8]) -> io::Result<()> {
+    pub(crate) fn put(&mut self, key: u128, bytes: &[u8]) -> io::Result<()> {
         let previous = self.index.get(key)?;
         let id = self.records.insert(bytes)?;
         self.index.insert(key, id.0)?;
@@ -71,14 +71,14 @@ impl PagedStore {
         Ok(())
     }
 
-    pub(crate) fn get(&self, key: u64) -> io::Result<Option<Vec<u8>>> {
+    pub(crate) fn get(&self, key: u128) -> io::Result<Option<Vec<u8>>> {
         match self.index.get(key)? {
             Some(id) => self.records.read(RecordId(id)),
             None => Ok(None),
         }
     }
 
-    pub(crate) fn delete(&mut self, key: u64) -> io::Result<bool> {
+    pub(crate) fn delete(&mut self, key: u128) -> io::Result<bool> {
         let Some(id) = self.index.get(key)? else { return Ok(false) };
         self.index.remove(key)?;
         self.records.delete(RecordId(id))?;
@@ -104,7 +104,7 @@ mod tests {
     fn rec(i: u64) -> Vec<u8> {
         format!("{{\"_key\":\"n{i}\",\"name\":\"record {i} west java\",\"n\":{i}}}").into_bytes()
     }
-    fn key(i: u64) -> u64 { i.wrapping_mul(0x9E37_79B9_7F4A_7C15) }
+    fn key(i: u64) -> u128 { i.wrapping_mul(0x9E37_79B9_7F4A_7C15) as u128 }
 
     #[test]
     fn records_round_trip_through_the_index() {
